% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Araw}
\alias{Araw}
\title{Raw acceleration}
\format{A matrix with three columns, corresponding to raw acceleration
values in the x-, y-, and z-axes. Includes an attribute: \code{Afs}, the raw
sampling frequency (400 Hz).}
\usage{
Araw
}
\description{
Raw acceleration data from Cade et al. 2018 JEB, collected at 400 Hz.
}
\keyword{datasets}
