% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prh}
\alias{prh}
\title{Pitch-roll-heading data.}
\format{A data frame with 8 variables: \code{time}, \code{depth},
\code{pitch}, \code{roll}, \code{head}, \code{Aw}, \code{Mw}, and \code{Gw}.
\code{Aw}, \code{Mw}, and \code{Gw} are matrix columns, with three-axis
acceleration, magnetometer, and gyroscope data, respectively, rotated into
the whale's frame of reference. Includes an attribute: \code{fs},
the downsampled sampling frequency (10 Hz).}
\usage{
prh
}
\description{
Example PRH data from Cade et al. 2018 JEB, downsampled to 10 Hz.
}
\keyword{datasets}
